inject: force sending a probe when receiving a mapping
This fixes a critical bug where when reciveing a new mapping, autopunch
would update its last_send to the current time, causing the relay loop
to not send any outgoing probe to its peer.

Instead, we set it to a special value that tells relay to send a probe
on its next loop.